backend/requirements.txt
pydantic
awesome-slugify
fastapi[all]
uvicorn
motor
git+https://github.com/codello/Motor-ODM.git